refactor: useProblemList 훅 분리
